+++ addon/plugins/article-structure-plugin/structures/title.ts
@@ -75,8 +75,10 @@ export const titleSpec: StructureSpec = {
 export const title = constructStructureNodeSpec({
   type: SAY('Title'),
   content: 'structure_header title_body',
+  allowSplitByTable: false,
 });
 
 export const title_body = constructStructureBodyNodeSpec({
   content: '(chapter|block)+|(article|block)+',
+  allowSplitByTable: false,
 });